Correct numbering in traditionalKnowledge.rst and spatialCoverages.rst
Numbering in traditionalKnowledge.rst was incorrect, needed to add a block that encompassed both Labels and Notices, then renumber each, ultimately adding a third level in the hierarchy to accommodate the addition of elements withing Labels and Notices. Also swapped numbering on spatialCoverages and TK pages, so that spatialCoverages comes first.