Update traditionalKnowledge to point to Local Contexts Hub Projects
Local Contexts Hub contains projects with validated TK/BC Labels and Notices. We point to the URI of the Projects in the Hub as the schema governing TK/BC Labels and Notices, then allow entry of a Project URI, which can be resovlved to retrieve the validated Labels and Notices.